+++ xenon-common/src/main/java/com/vmware/xenon/services/common/authn/AuthenticationRequest.java
@@ -19,5 +19,5 @@ public class AuthenticationRequest {
         LOGIN, LOGOUT
     }
 
-    AuthenticationRequestType requestType;
+    public AuthenticationRequestType requestType;
 }